agents/pyproject.toml
[project]
name = "adk-samples"
version = "0.1.0"
description = "Repository of different ADK (Agent Development Kit) agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "opik",
    "exa-py",
    "pydantic>=2.0.0",
    "python-dotenv",
    "asyncpg",
    "greenlet>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

